Refactor project layout for upcoming extension
I will be extending the codebase today. This is some refactoring
to clean it up before extending functionality and test suite.

Not abstract/yak-shaving - Locally I've added stub modules for what's
coming up next, and it's helped sketch a sensible layout.
(Stub/dummies not committed, as they are not active code.)