Add special treatment for sentences to _crude !
_crude was missing that special sauce for how to treat sentence-beginning specially.
I waited a bit before adding that, did other things, as I didn't want to 'cheat',
wanted to understand it exactly.

Came back with fresh eyes and it was very clear what to do, so I've added it.

Also added some mess of comments,
for other changes to make here to improve _crude.