Fix critical bug in SanitizedString! (Keep newlines)
Forgot for a bit that newlines and carriage returns would be removed by this,
really I was trying to handle null bytes and so on.

but of course, of course, CRLF injection, yes, those are control characters too,
just tricked myself for a bit and sent myself down a debugging trail...
but now sorted out.