Dealing with legacy PyMarkovChain quirks further
- get it to play nice with the other implementation(s)
in here - i.e. stop the token-joining, we've separated
that concern out to the callers, the Markov Chain model
should just be returning sentences (as lists of lists,
i.e. lists of <lists of words>)

- littering some deprecation warnings in there,
I am going to take an axe to it soon

Cleanup to pymc aka PyMarkovChain tests

----

squshed:
Fix a regression in PyMarkovChain module
from all this hack and slash